+++ createDirectoriesTree.sh
@@ -4,13 +4,13 @@
 log "#### Creating directories tree ####"
 
 cd ~/ &&
-mkdir .gradle ||
-mkdir Coding ||
+mkdir .gradle
+mkdir Coding
 cd Coding &&
-mkdir IDEs ||
-mkdir xDKs ||
-mkdir servers ||
-mkdir tools ||
-mkdir workspace-mine ||
-mkdir workspace-test ||
+mkdir IDEs
+mkdir xDKs
+mkdir servers
+mkdir tools
+mkdir workspace-mine
+mkdir workspace-test
 mkdir workspace-work