+++ modules/tomcat.sh
@@ -7,7 +7,7 @@ preInstallationLog "Apache Tomcat"
 
 enterDirOrExit ~/Coding/servers/
 
-TOMCAT_DOWNLOAD_URL=http://apache.dattatec.com/tomcat/tomcat-8/v8.0.33/bin/apache-tomcat-8.0.33.zip
+TOMCAT_DOWNLOAD_URL=http://apache.dattatec.com/tomcat/tomcat-9/v9.0.20/bin/apache-tomcat-9.0.20.zip
 downloadAndUncompress "Apache Tomcat" "$TOMCAT_DOWNLOAD_URL"
 
 TOMCAT_DIRECTORY=$(getFileNameWithoutExtension "$(getFileName "$TOMCAT_DOWNLOAD_URL")")
@@ -17,3 +17,5 @@ chmod +x catalina.sh &&
 
 postInstallationLog "Apache Tomcat"
 logInfo "Apache Tomcat configured successfully. Now you can run: 'sh startup.sh' and 'sh shutdown.sh'"
+
+enterDirOrExit "${CURRENT_DIR}"