+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 ## Removed
 - Deprecated links and graphics from README.
 - Dex2Jar Android utility in favor of user specific needs.
+- Go language in favor of user's specific needs.
 
 # v2.2.0
 ## Added